Move get searchable fields to index class
