+++ packages/seal-algolia-adapter/AlgoliaSchemaManager.php
@@ -4,7 +4,6 @@ namespace Schranz\Search\SEAL\Adapter\Algolia;
 
 use Algolia\AlgoliaSearch\SearchClient;
 use Schranz\Search\SEAL\Adapter\SchemaManagerInterface;
-use Schranz\Search\SEAL\Schema\Field;
 use Schranz\Search\SEAL\Schema\Index;
 use Schranz\Search\SEAL\Task\AsyncTask;
 use Schranz\Search\SEAL\Task\TaskInterface;
@@ -42,7 +41,10 @@ final class AlgoliaSchemaManager implements SchemaManagerInterface
     {
         $searchIndex = $this->client->initIndex($index->name);
 
-        $attributes = $this->getAttributes($index->fields);
+        $attributes = [
+            'searchableAttributes' => $index->searchableFields,
+            'attributesForFaceting' => $index->filterableFields,
+        ];
 
         $indexResponse = $searchIndex->setSettings($attributes);
 
@@ -53,53 +55,5 @@ final class AlgoliaSchemaManager implements SchemaManagerInterface
         return new AsyncTask(function() use ($indexResponse) {
             $indexResponse->wait();
         });
-    }
-
-    /**
-     * @param Field\AbstractField[] $fields
-     *
-     * @return array{
-     *     searchableAttributes: array<string>,
-     *     attributesForFaceting: array<string>,
-     * }
-     */
-    private function getAttributes(array $fields): array
-    {
-        $attributes = [
-            'searchableAttributes' => [],
-            'attributesForFaceting' => [],
-        ];
-
-        foreach ($fields as $name => $field) {
-            if ($field instanceof Field\ObjectField) {
-                foreach ($this->getAttributes($field->fields) as $attributeType => $fieldNames) {
-                    foreach ($fieldNames as $fieldName) {
-                        $attributes[$attributeType][] = $name . '.' . $fieldName;
-                    }
-                }
-
-                continue;
-            } elseif ($field instanceof Field\TypedField) {
-                foreach ($field->types as $type => $fields) {
-                    foreach ($this->getAttributes($fields) as $attributeType => $fieldNames) {
-                        foreach ($fieldNames as $fieldName) {
-                            $attributes[$attributeType][] = $name . '.' . $type . '.' . $fieldName;
-                        }
-                    }
-                }
-
-                continue;
-            }
-
-            if ($field->searchable) {
-                $attributes['searchableAttributes'][] = $name;
-            }
-
-            if ($field->filterable) {
-                $attributes['attributesForFaceting'][] = $name;
-            }
-        }
-
-        return $attributes;
     }
 }
